Work around lack of default values for fields
I have to allow 'unlock' to be null, becuase I can't set a default value with Jaguar (yet) and the CMS does not supply an unlocked value.

Have implemented a getter that will return false if 'unlock' is null, to work around this.

If the unlock code is not set, the quiz will always return as unlocked.